Upgrade gradle and project dependencies
This commit also switches to maven-publish since bintray is not a viable
option anymore. Note: metadata is left commented out in a few places
because this information will be useful in later commits that
specifically address publication requirements. The primary focus for now
is to update / upgrade the APIs used in this implementation. It can be
tested with other projects via `mavenLocal()`.